Lets namespaces be created/reconciled out-of-order
Previously, if we hadn't actually seen a namespace on the apiserver, we
wouldn't create an entry for it in the forest. With this change, we
create an entry in the forest for any *referenced* namespace, even if it
doesn't exist, and mark that ns as existing or not existing as
appropriate. This allows us to catch state changes (ie, not existing to
existing) and therefore to update any other affected namespace.

More concretely, this allows namespaces to be created out of order, and
as a side effect elegantly solves reconciliation problems when the
controller is started (since the controller reconciles namespaces in an
arbitrary order). If a child is reconciled first, it may briefly be
given the "missing parent" condition, but when the parent is later
reconciled, it will notice that the child needs to be reconciled too,
and the condition will then be removed.

Tested: new integ test, manual