Restructure NS enqueuing; improve thread safety
Rather than collecting a list of namespaces to mark for enqueueing after
`syncWithForest` is complete, this change allows them to be enqueued
inline, which simplifies the control flow.

This also means that it's possible for the same namespace to be
reconciled twice at the same time (it was *probably* possible before as
well). While syncWithForest is guarded by the forest-wide mutex, there
is a chance that the results could be committed to the apiserver
out-of-order. I don't *believe* I ever observed this (a bug I was
chasing turned out to be a more mundane logic error) but in order to fix
this potential problem, this change introduces a per-namespace lock that
is only released after the namespace has been _fully_ committed.

This allows multiple namespaces to be reconciled in parallel (although
access to the forest is still globally synchronized), but prevents any
*one* namespace from being reconciled more than once at a time.

Tested: integ tests; manually with later changes