Implement self-service namespace creation
This change notices required children that are missing and ensures they
get created. This is the first time we also modify namespaces, as
opposed to simply hierarchy singletons.

This change does *not* include a validating webhook to prevent most
failure cases that are identified by the conditions; this will come in a
future commit. It also doesn't attempt to set up ownersReferences to
automatically delete child namespaces when parents are deleted, although
we should probably do that too.

Tested: manual plus new integ tests