Obey generic CLI flags
This change allows kubectl-hnc to respect the `--as` flag (among other
standard flags), which is essential for testing HNC features that depend
on the identity of the caller, such as our planned validation
controllers).

Tested: granted various service accounts get+update rights on namespaces
and hierarchyconfigurations at various points in a complex namespace
tree. Verified that `kubectl hnc --as <sa> tree <ns>` worked or was
denied as expected.