Check authz when changing a namespace's parent
This implements the logic described in the design doc to prevent a
subadmin from depriving a superadmin of access to a namespace by
assigning it away, and from gaining access to a new namespace's
propagated objects by joining it as a child.

Tested: new unit tests. Also manually verified that by default, acting
as a service account did not allow me to change the hierarchy of a
subnamespace even if I had get/update permissions on its
hierarchyconfig. Verified that as I added rolebindings to the parents,
the error messages changed accordingly as the reasons for failures
changed until I had sufficient permissions, at which point it passed. I
have not tested this exhaustively, apart from the unit tests.