Set object conditions to replace modified annotation
This replaces modified annotation with object conditions when object is
detected as modified from source object. It will set ObjectOverridden
Condition on the modified object and ObjectDescendantOverridden
Condition on the source object. Before propagation, if any ancestor of
the object has ObjectOverridden Condition, it will not be propagated.

Tested: make test and manual test in Kind.