Change API package import to 'api'
We used to import the HNC API as `tenancy` because the APIs were
initially in the `tenancy.k8s.io` group (or something like that). Now,
however, they're in `hnc.x-k8s.io` and calling that API "tenancy"
doesn't really make sense anymore. Importing it as `api` is unambiguous
in the context of the HNC (all other imported APIs are named) and is
also shorter to type and easier to read.