Enable multithreading in the HC controller
Tested: manual testing on GKE. I also instrumented the logs (removed in
this commit) to report the maximum actual concurrency when I started 100
threads; it peaked at about 15. However, the "startup time" (the time
from when the controller starts to when it stops making changes) didn't
change on my test cases and remained constant at about 30s. However,
think it's a good idea to turn this on now to flush out any problems.