HNC: forbid propagated objects from being modified
Manual test:
org1
|
- team1

(1) Create a secret in org1. Modification to the origin secret is
allowed, but modiciation to the secret in team1 is not allowed.
(2) Create ResourceQuota in org1. Create a pod in org1 and see
ResourceQuota has been partially used, but the propogated ResourceQuota
for team1 is not affected.