Give the HNC service account all required verbs
I noticed that HNC wasn't able to copy a rolebinding to cluster-admin
since it was missing at least two verb (delete-collection and
impersonate). I added both of those as well as "use" (which, AFAIK, is
only used for PSP which aren't namespaced but it seemed like a safe
addition) and verified that HNC was now able to copy the rolebindings.

Since cluster-admin is the most powerful role in K8s (again, AFAIK), I
think this means that HNC can now copy any rolebinding.