Store CritAncestor condition in the forest
CritAncestor condition is set when any ancestor has local critical
conditions. Currently, it's only set in the HierarchyConfiguration
object. This commit sets the CritAncestor condition in the forest first
with ancestor namespace as the key. This change enables object
controller to just look in the forest to find all critical conditions in
the namespaces, so that it can stop propagation if any condition
detected.

Tested with "make test" and "make test NOC=1" and all tests passed.