Check Crit condition in object controller
This commit stops the object controller for further action if any
critical condition is found in the namespace. The object controller will
propagate/delete objects after the critical conditions are gone.

Tested with "make test NOC=1" with additional tests. It passed both
"make test" and "make test NOC=1". When commenting out the "create
missing parent" code in the test and some of the following expectations,
the test failed at expectations that we expect it to fail.