Improve kubectl subcommands to match naming doc
See bit.ly/hnc-cli-naming for details. The two major changes here are
to rename `show` to `describe`, and to add a `create` subcommand as a
synonym for `set --requiredChild`.

In addition, this change allows both `set --requiredChild` and `set
--optionalChild` to take multiple values, either comma-delimited or with
repeated flags. It also slightly improves error messages.

Tested:
* Tried both --requiredChild and --optionalChild with a combination of
comma-delimited and repeated flags.
* Manually tried the `create` command.
* Manually tried the `describe` command.