Create reconcile counters and a flag to enable it
This commit created reconcile counters to get total/current number of
hier/object reconciliation attemps. It generates logs every 0.5s when
the controllers are working. Since the logs are only for performance
testing, a flag is set to enable/disable the counter & the logging.

Manually tested on GKE cluster by filtering the logs.