Improve RequiredChildConflict condition
This change downgrades RequiredChildConflict to be a non-critical
condition. It also only annotates it on the namespace *requesting* the
required child, not the child itself, so that one tenant cannot cause
another tenant's namespace to be marked as having a condition (critical
or otherwise).

Testing: manual with the validating webhook disabled. Found one bug and
added the unit test.