Remove redundant per-namespace locks
Also removes some verbose logs from the new object controller.

Tested: this one's hard to test. I manually created a namespace with 10
children and restarted the controller, and the root is only reconciled
twice even though each of the children try to enqueue it. Then again, we
barely testing multithreading before enabling it and the client code
referenced in
https://github.com/kubernetes-sigs/multi-tenancy/issues/212 seems pretty
clear, so I think this is fairly safe.