Fully automated builds with all binaries on Cloud Build
This change updates `make release` so that as little as possible happens
on the host computer (only the manifests are built locally, and then
only because I haven't figured out how to get kustomize to run on Cloud
Build yet). Both the plugin and the image are now build on Cloud Build,
with nothing touching the builder's workstation - all code is pulled
directly from Github.

The image does still have to make a round-trip onto the builder's
workstation since the Cloud Build service account doesn't have
permission to push directly to the K8s staging GCR, but you can still
verify that the image built on Cloud Build is the same one on the
official GCR.

Tested: followed these instructions to create a new build and deployed
it.