Use Cloud Build for everything and update docs
This change updates `make release` so that as little as possible happens
on the host computer. All artifacts - the manifests, kubectl plugin and
of course the image itself - are now built on Cloud Build.

The image does still have to make a round-trip onto the builder's
workstation since the Cloud Build service account doesn't have
permission to push directly to the K8s staging GCR, but you can still
verify that the image built on Cloud Build is the same one on the
official GCR.

Tested: followed these instructions to create a new build and deployed
it.