Propagate empty source objects to descendants.
Currently an empty source object will not be propagated to descendants if the descendants do not have any copy of the object as reported in #441. The PR fixes this issue so that empty source objects will be propagated to descendants if the descendants do not have the copy of the object.