Create empty HierarchicalNamespaceReconciler
Create the empty HierarchicalNamespaceReconciler to only generate a log
with the information of the object being reconciled and setup with
controller manager.

Add "forbidden" state to CRD status.
Rename HierarchyReconciler to HierarchyConfigReconciler.

Tested on GKE: I created a sample hns object and found the log of
reconciling that hns object in Stackdriver.