Update HNS and HC reconcilers to allow hns creation
Create namespace and hierarchyconfig instances of the self-serve
subnamespace in the HNS reconciler.

Make changes to the HC reconciler to use the list of hns objects in the
namespace as requiredChildren instead of using the
spec.requiredChildren.

Tested on GKE cluster. First, added the
'enable-hierarchicalnamespace-reconciler` flag into the
config/default/manager_auth-proxy_patch.yaml file. Used 'kubectl hns
create2 -n parent child' and found the hc and the hns objects were
created in the 'parent' namespace. 'child' namespace and its hc object
were also created. I also saw the test log in Stackdriver saying that
the 'requiredChildOf' field of the subnamespace in the forest was set to
the 'parent.`