Support 'ignore' mode in ConfigReconciler
This PR adds the support for 'ignore' mode in ConfigReconciler. Specifically, it adds following features:

If the mode of a type is set to 'ignore', the corresponding object reconciler will ignore all subsequent reconciliations (e.g., new or changed objects will not be propagated).
If the mode of a type is changed from 'ignore' to 'propagate', reconciliations for all existing objects of the type in the cluster will be triggered.
If a mode of a type is unrecognized, it will be treated as 'ignore'.
If a mode of a type is unset, it will be treated as 'propagate'.
Tested: GKE cluster; unit tests.

This PR includes changes introduced in #462

Design doc: http://bit.ly/hnc-type-configuration
Issue: #411