Update `isModified()` to remove sleeps in tests
Update `isModified()` to return a function for `Eventually` to retry.
Change `ExpectWithOffset` to `EventuallyWithOffset` to get the object.
This prevents wierd cases of failure when the object does exist but the
first attempts of getting the object fails.

Tested with "make test" for 20+ times to make sure it's not flaky.