Set an object reconciler to "ignore" mode for a removed type
If a type is removed from the HNCConfiguration Spec, we will set the corresponding object reconciler to "ignore" mode.

Ideally, we should shut down the corresponding object reconciler. Gracefully terminating an object reconciler is still under
development (https://github.com/kubernetes-sigs/controller-runtime/issues/764). Once the feature is released, we will see if we can shut down the object reconciler instead of setting it to "ignore" mode.