Add better HNC Config observability and fix flaky test
One of the HNC config observability tests was failing because we didn't
wait for the HNC config reconciler to pick up a change before verifying
that it was working:

* We changed a sync mode from "propagate" to "ignore"
* We created a new object
* We verified that this object wasn't created

Occasionally (about 1/4 of the time), the object reconciler would run before
the HNC config reconciler had a chance to change it to "ignore." The fix
is to add a "Mode" to the HNC Config status, so we can tell when the
reconciler has actually picked it up. This is also useful for when the
spec doesn't include an explicit mode.

This change also includes a few other minor improvements:

* If there's an extra HNC Config object, and we can't write it back due
to an apiserver error, retry the failure (previously we wouldn't retry).
* Fully deterministically sort the types in the Status (had only been
sorting by apiVersion)
* Added comments throughout
* Make the Gomega pattern matching more explicit for easier-to-read
tests and better error messages (e.g. when the sync mode is wrong, the
bad value will now be printed out)
* Other minor test improvements

Tested: ran `make test` 10x