Always sync objects to forest, even in namespaces with critical conditions
See #347. When an object is removed, we need to remove it from the
in-memory forest even when the object is in a namespace with a critical
condition, because when the condition is resolved, we'll use the forest
to re-sync all descendants of that namespace, and if the forest has
stale objects in it, the propagated copies won't be removed.

Added several other cleanups including:

* Made several names more unique so they can be more easily searched
(e.g. replaced `delete()` with `deleteObject()`)
* Factored out a new syncMissingObject() functionout of Reconcile()
* Some cosmetic renames plus a bunch of longer comments

TESTED: new integ test fails without the fix and works with it; also
tested manually on GKE.