Replace ns.Owner with ns.IsOwned in the forest
Since we only use subnamespace owner annotation as source of truth and
always set parent to the owner in HCR, ns.Owner string field is not
necessary and a ns.IsOwned bool field is enough.

Tested with 'make test' and manually tested the webhook of changing an
owned namespace's parent on GKE cluster.