Flush obsolete conditions when tree changes (#328)
This changes ensures that any conditions on ancestor namespaces caused
by objects in their descendants are removed if a child namespace is
removed from its hierarchy. Without this change, no reconciler will ever
attempt to remove the CANNOT_PROPAGATE_OBJECT condition from an
ancestor, and it will incorrectly sit on the (old) ancestor namespace
until HNC is restarted.

I initially tried to fix this by having the object reconciler determine
from scratch whether it was needed, by searching the forest for all
possible copies of the source object and looking for conditions. But I
found that approach to be overly confusing and it still needed to be
triggered in a variety of ways.

I don't really love the fact that most of the code here is in
hierarchy_config.go; it would be more natural to put it in object.go.
But I'm not sure exactly how I'd trigger it. For now, I'd rather simply
submit this fix, and then make incremental improvements later if anyone
gets any bright ideas.

Tested: hack/test-issue-328.sh. I wasn't able to induce this kind of
situation in the integ tests, because HNC has full permissions in those
tests, unlike on a real cluster where it runs with its own more
restrictive service account. The new script fails without this fix and
passes with it; it works even better when #326 is fixed as well, but I'm
submitting that in a separate PR.