Test if obsolete CannotUpdate condition gets cleared in descendants
We suspect obsolete CannotUpdate condition will be cleared on the child
namespace who just changes its parent but not on its descendants.
Therefore, this new test case is added to verify that.

It turned out the condition was cleared eventually on the descendants.
However, the object reconciliation was triggered by controller-runtime,
not by HNC reconcilers. Thus this test is tailored to be able to tell if
the object enqueue is triggered by controller-runtime or our HNC
reconcilers. The difference is that controller-runtime has exponential
backoff. After issue 605 is fixed, we should see the obsolete conditions
in descendants are gone immediately instead of waiting for the
controller-runtime exponential backoff retry.

Tested by running this script and could see around first 9 out of 10
attempts still with conditions but evetually cleared at the 10th
attempt for now (before issue 605 is fixed).