Don't create HC in namespaces being deleted
I noticed while doing a cascading deletion that the HCR was trying to
write back the HC, because after it's deleted, it looks like we want to
sync it with the namespace's owner attribute. This was causing a lot of
errors in the log because you can't create an object in a namespace
that's being deleted. This change allows HNC to continue writing back
_existing_ HCs (e.g. to remove their finalizers) but will never write
back a _new_ HC to a namespace being deleted.

I also took out the code that allows the HCR to _create_ namespaces
since this is now owned entirely by the HNSR.

Tested: watched logs during cascading deletion before and after this
change. Before, there were tons of error messages (although it did
successfully delete the NS); after, the log was clean.