Better flags, output and examples for hns-set
The main change here is to add --forbidCascadingDelete, as well as some
short flags (-p for --parent, -r for --root, -a for
--allowCascadingDelete and -f for --forbidCascadingDelete). It also adds
an "examples" section to the documentation and prints out fewer messages
if allowCD is unchanged.

Tested: manual testing for all the modified variants of 'set'; verified
that new example documentation looks good.