Flush obsolete conditions on descendants
As described in issue 605, obsolete CannotUpdate condition was not
cleared on the descendants namespaces of a namespace who just changes
its parent. This commit changes the key of a CannotUpdate condition from
the local object to the source object if it exists. Therefore, we can
flush obsolete conditions on descendants by clearing the conditions by
ancestor namespaces, the same way as how we flush obsolete conditions on
ancestors.

Tested with "make test" and running the test script
"hack/test-issue-605.sh' seeing the CannotUpdate condition was gone
immediately in all 10 attempts.