Make HNC container run as non root
Make the HNC container image run as non root. Add securityContext in the
HNC pod to make containers in it run as non root.

Add a port flag that defaults to 443 in the code. Since the container is
run as non root, it's not able to use ports below 1024 reserved for root
only. Therefore, in the pod manifest, set port flag to 9443 and have the
webhook service map the 443 port to targetPort 9443 in the conainer.

Tested on a GKE cluster with PSP of notrunasroot is enabled.