More robust server checks
Now that we have subnamespaces, we can't have missing namespaces in the
middle of a tree, only at the root. Simplify the authz checks to reflect
this, and better handle the two reasons why the namespace might be
missing (HNC is starting up vs the namespace is truly orphaned).

Tested: manually confirmed that we can fix namespaces with the
CritParentMissing condition, either by making it a root or recreating
the missing parent. Also manually confirmed that we can't change the
properties of a namespace with a critical ancestor. Added logs and
verified that we're checking authz on the correct namespaces in multiple
cases. Also added new tests to confirm all of this automatically.