Stable removal of obsolete conditions
I never liked the fix to #328 since there was basically only one chance
to notice that conditions were now obsolete - when the parent was being
changed. This felt weird, and when I started looking at (fittingly)
issue #666, which requires us to allow cycles in the in-memory forest,
it became difficult to see how to make the existing logic work.

This new approach, by contrast, annotates each condition code with the
knowledge of whether it can be caused by an ancestor or descendant
namespace, making it easy to look at the *current* state of the forest
(and not of any one change) and determine whether the condition should
be removed. This will make future logic changes easier and just seems
cleaner overall.

This also makes a minor improvement to kubectl-hns-describe. It also
removes the AffectedObject from the HNSMissing condition, because it's
not really necessary (it always points to the immediate parent) and
because it made the logic significantly easier.

Finally, I noticed that markExisting() wouldn't always do the right
thing for subnamespaces because it's called before syncParent(), so I
fixed this. I doubt this was causing any problem (I think we always
re-sync a namespace at least once) but we may as well be accurate.

Tested: re-ran hack/test-issue-328.sh (it worked). Played around a bit
by hand and noticed that everything seemed to be working well.