Better handle namespace cycles
If a user bypasses the admission controller, they can create namespace
ancestry cycles, but these have never been correctly represented
in-memory in HNC, and therefore were not correctly reported. For
example, only _one_ namespace in a cycle would have the
CritInvalidParent condition, as described in #666.

This commit includes several tightly-coupled changes to fix this:
* The Forest no longer prevents you from creating cycles. Instead,
Namespace.CycleNames() returns a non-nil list of a namespace is in a
cycle, and other methods like AncestryNames() and DescendantNames() are
now cycle-safe.
* The kubectl-hns plugin correctly handles cycles
* The HC reconciler ensures that other cycle members are enqueued when a
cycle is created or broken, and uses the new CritCycle condition on all
members of the cycle instead of the CritInvalidParent condition on only
a single member.

Tested: updated various unit tests. Also verified the following:
* With the admission controller disabled, I can create cycles. All
members of the cycle get the CritCycle condition annotated or removed as
appropriate.
* Descendants of namespaces in a cycle continue to get the CritAncestor
condition.
* With the admission controller *enabled*, I can resolve cycles. I
didn't test anything fancy with the authz checks as the recommended mode
to solve these problems is "I have admin so I can do anything," but the
new unit tests do verify that you need authz to break a cycle.