Add namespace conditions to HNCConfig and metrics
This change adds a new namespaceConditions field to HNCConfig, as well
as as populating a namespace_conditions metric tagged by both the
condition as well as whether or not it's critical. This allows admins to
set up monitoring for bad behaviour.

Tested: induced the CritParentMissing condition and observed that all
affected namespaces were listed in HNCConfig, and that the metric counts
and tags were correct in Stackdriver.