Allow namespace deletion in K8s 1.14
See https://github.com/kubernetes-sigs/multi-tenancy/issues/688. If we
can't validate the deletion because OldObject doesn't exist, we need to
just allow the deletion.

In the case of a namespace with children and allowCascadingDelete
*unset*, this will delete the parent but *not* the children; they will
be left with two conditions (CritMissingParent and
MissingSubnamespaceAnchor). If allowCascadingDelete *is* set, everything
will work properly.

Tested: verified that everything works as expected in 1.14 and that
there are no changes in 1.16.