Reconcile external namespaces and update tree labels
Add forest.Namespace.ExternalTreeLabels field to store external
hierarchy in memory. Add syncExternalNamespace() func in
HierarchyConfigReconciler to update the external tree labels in the
forest. Enqueue descedants if the namespace is converted from external
to internal or vice versa so that the external tree labels can be
removed or propagated to the descendant namespaces.

Early exit syncSubnamespaceParent(), syncParent(), syncLabel() for
external namespaces. In syncLabel(), for an internal namespace, if the
root ancestor is an external namespace, propagate all the external tree
labels too.

Tested by "make test" with three new integration test cases.