Don't delete subnses with the Anchor CRD
A reasonable admin action is to call `kubectl delete -f manifests.yaml`,
but this will delete the CRDs and hence the subnamespace anchors. If any
namespace have cascading deletion enabled, this will delete the
namespaces themselves, which is surprising and destructive. This change
disables cascading deletion if the Anchor CRD itself is being deleted,
as opposed to any one anchor CR.

Tested: Added hack/test-delete-anchor-crd.sh and confirmed that it fails
on HEAD and passes with this fix. Also verified that log messages looked
good. Tried deleting some subnamespaces with and without setting
allowCascadingDelete and everything worked properly. Also reran
test-issue-501.sh, which _mostly_ worked although it looks like the
webhooks now prevent us from getting a SubnamespaceAnchorMissing
condition (which is good).