Disable leader election
We only have one pod in our deployment, so leader election does nothing
but add about 10s of delay when HNC starts up. Let's get rid of it for
now.

Tested: measured ~11s of startup delay with and without leader election.
I've been running without it for a few days and everything seems to work
fine.