Don't unpropagate objects after CRDs are deleted
This change ensures that if the HierarchyConfig CRD is deleted before
the HNC pod, it will not delete the propagated objects. This reduces the
surprises when uninstalling HNC.

Tested: verified by hand that I can still delete the HC and this has the
effect of resetting the parent, and that cascading deletion still works.
Wrote a new test that verifies that the objects are not deleted after
five seconds, and verified that this test fails on HEAD but passes with
this patch.