Let HNC propagate builtin admin rolebindings
I _think_ HNC used to be able to propagate the `admin` role, but there
were several permission (deletecollection, impersonate) that don't
appear to be working in 1.15. I've now added these permissions to the
HNC service account so it can grant them.

Tested: new hack/test-issue-772.sh verifies that HNC can propagate the
`admin` builtin; it fails without this change and passes with it. The
existing hack/test-issue-771.sh, which uses `cluster-admin`, continues
to be unpropagateable as expected.