Enable object webhooks dynamically based on hncconfig
Remove object webhooks of different types in the hardcoded list.
Instead, have one objects webhook for all types and in the Handle()
function, dynamically look up the current type configuration in the
forest and allows changes on all other types that are not in propagate
mode in the type configuration. As long as the request object has the
matching GroupKind in propagate mode in type configuration, we don't
allow changes to it.

Tested by newly added unit test cases and manually configuring types and
modifying objects.