Enable removing bad subns anchors safely
Allow removing bad subns anchor if the subnamespace already exists and
the parent doesn't match. The race condition could happen when two
parents both create the same subns the same time and the webhook is
bypassed because the forest is not updated after the first change. We
would always allow resolving a bad state.

Most importantly, make sure the subnamespace is not deleted with the
deletion of a bad anchor. Before this change, the subnamespace created
from a good anchor will also be deleted when deleting a bad anchor.

Tested by test-issue-797-leaf.sh and test-issue-797-non-leaf.sh, where
they disable the webhook first to generate the race condition in both a
leaf and non-leaf subnamespace and enable the webhook to test the
change. One unit test case is added to the anchor validator too.