Set apiVersion and kind as required fields in hncconfig types
apiVersion and kind fields in hncconfig.spec.types should never be
empty. It was a mistake to leave them omitempty. The current validating
webhook ensures the type generated from these two fields are valid,
therefore, empty fields are not allowed. Still we need to fix it in the
CRD.

The apiVersion and kind fields in hncconfig.status.types are also
updated.

Tested with the webhook disabled and it was still not allowed.