Allow finalizers to be removed when deleting CRDs
See #824. When the CRDs are being deleted, we need to continue
reconciling the CRs until they're gone in order to ensure that any
finalizers on the CRs are removed.

Tested: new test script usually fails without this fix, but reliably
passes with it. Observed the new log messages working as expected while
running the script. For subnamespace anchors, tried removing the
CRD-checking code entirely and verified that
hack/test-delete-anchor-crd.sh failed, and then replaced it with the
current code (the shared isDeletingCRD function) and verified that it
passed again.